requirements.txt
# Core
python-dotenv>=1.0.0

# Notion
notion-client>=2.2.1

# LangChain
langchain>=0.3.0
langchain-openai>=0.2.0
langchain-community>=0.3.0
langchain-chroma>=0.1.0
langchain-text-splitters>=0.3.0

# Vector store
chromadb>=0.4.0

# LLM & embeddings
openai>=1.0.0

# Chat UI + Google auth
streamlit>=1.42.0
streamlit[auth]>=1.42.0
pyyaml>=6.0
